requirements.txt
mcp[server]>=1.6.0
libvirt-python>=8.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
pytest-cov>=4.1.0 